ic: migrate off src/cmd/cc onto compilers/cck
Same migration as 7c and vc (5e1c604e2, 4346c480a), same core shape:
repoint compilers/ic/mkfile's HFILES/LIB and compilers/ic/gc.h's
#include at compilers/cck instead of src/cmd/cc, then fix the same
small set of incidental mismatches -- outstring/gextern/exreg/ncast's
int32 params/return changed to long, xround renamed to round, and
p->reg = textflag reverted to p->reg = (profileflg ? 0 : NOPROF)
(dead Go-runtime-era cruft 5c/6c/8c had already dropped; ic/txt.c's
own "//goken:" comment already pointed at the fix).

ic needed one thing beyond that shared set: compilers/cck/lex.c's
main() never had the argv[0]-basename check
(`if(*p == 'j') thechar = 'j';`) that lets the ic/jc riscv32/riscv64
dual-width binary tell which name it was invoked under -- src/cmd/cc/
lex.c already had it (ic's previous frontend), but compilers/cck/lex.c
never needed it before since none of its other consumers (6c/5ck/8ck)
share a binary under two names. Without it, `jc` silently produced
.i-suffixed object files instead of .j. Ported the check verbatim from
src/cmd/cc/lex.c into compilers/cck/lex.c, placed before ginit() as
before -- confirmed safe for 5ck/6c/8ck since their own ginit()
functions unconditionally overwrite thechar regardless of this
pre-check. Also fixed ic/txt.c's riscv64 dual-width table copy,
`memmove(ncast, ncast64, NTYPE*sizeof(int32))`, to sizeof(long) now
that ncast/ncast64 are long -- otherwise it would silently copy only
half the table.

Verified: clean build, tests/c/{mini,mini2,vlong} under qemu-riscv32,
tests/c/mini2 under qemu-riscv64 (jc; mini/vlong have no riscv64 .s
stubs at all -- a separate, pre-existing gap), and the full
tests/c/regressions suite (including riscv64_ptr_width, riscv64_sraiw,
float_const_oob_riscv/riscv64) all passing. benchs/compcert's
test_goken for mips was skipped for the same reason it was for vc:
known cross-arch issues there, not a useful signal yet.

This closes out todo.org's "migrate 7c/vc/ic off src/cmd/cc" item --
all three now build against compilers/cck. src/cmd/cc/ itself has no
remaining consumers among 5c/6c/7c/8c/vc/ic, though root mkfile's DIRS
still builds it (left as-is, flagged as a follow-up in todo.org).
Unifying compilers/cck/ itself into compilers/cc/ (needed for 6c/5ck/
8ck too, to fully match the stated end policy) remains separate,
larger, unstarted work.

+++ compilers/ic/gc.h
@@ -1,4 +1,4 @@
-#include	"../../src/cmd/cc/cc.h"
+#include	"../cck/cc.h"
 #include	<obj/i.out.h>
 
 /*
@@ -270,11 +270,11 @@ void	swit2(C1*, int, int32, Node*, Node*);
 void	casf(void);
 void	bitload(Node*, Node*, Node*, Node*, Node*);
 void	bitstore(Node*, Node*, Node*, Node*, Node*);
-int32	outstring(char*, int32);
+long	outstring(char*, long);
 int	mulcon(Node*, Node*);
 Multab*	mulcon0(int32);
 void	nullwarn(Node*, Node*);
-void	gextern(Sym*, Node*, int32, int32);
+void	gextern(Sym*, Node*, long, long);
 void	outcode(void);
 void	ieeedtod(Ieee*, double);
